pyproject.toml
[tool.poetry]
name = "project-barcode-to-img"
version = "0.1.0"
description = "Project to recognize a barcode from a camera"
authors = ["gustavoDwsantos <gustavos200116@gmail.com>"]
readme = "README.md"
packages = [{include = "project_barcode_to_img"}]

[tool.poetry.dependencies]
python = "^3.10"
pillow = "^9.5.0"
opencv-python = "^4.7.0.72"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
